updates shelving descrepency report to allow job number search, updates background queue service worker to properly pass error messages on failed requests, updates background service queue to reload when coming back online, updates service worker updates to refresh all service workers on new app updates

+++ src/layouts/MainLayout.vue
@@ -135,7 +135,9 @@ const checkForServiceWorkerUpdates = () => {
       // update the service workers and get latest content and refresh all indexDb instances
       const dbs = await window.indexedDB.databases()
       dbs.forEach(db => {
-        window.indexedDB.deleteDatabase(db.name)
+        if (db.name !== 'workbox-background-sync') {
+          window.indexedDB.deleteDatabase(db.name)
+        }
       })
       await registration.update()
     }
